+++ erudit/models/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
 from .core import *   # noqa
+from .grey_literature import *  # noqa
 from .journal import *  # noqa
 from .thesis import *  # noqa